add ClusterSecret
As per conversation with Lance, adding an additional layer of security
for KubeSpan is important to mitigate the honeypot of the central.
By requiring an additional encryption key which is _not_ published to a
registry to be able to connect to KubeSpan, we eliminate several types
of attack, namely:

 - peer poisoning by determining ClusterID
 - coordination server attack to gain access to Cluster IDs

This patch adds a wireguard-compatible Cluster Secret which is generated
along with the Cluster ID and which is used to further secure KubeSpan
wireguard connections.

Signed-off-by: Seán C McCord <ulexus@gmail.com>