refactor: get rid of prometheus/procfs dependency in pkg/resources
This PR introduces the concept of the resource adapter: something which
wraps core resource (which defines presentation) and adds additional
fatures to it which we don't want to have in `pkg/machinery` in the
future, but at the same time there's value in keeping it next to the
resource itself.

If the concept looks good, I will continue doing it same way for other
resources.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>