fix: unblock events watch on context cancel
Fixes #4653

This fixes an issue when `etcd` get stuck in `Pre` state so that even
`bootstrap` request doesn't kill it. This looks like a cluster which
fails to bootstrap itself and get stuck on reboot.

Even though `Watch()` aborts on cancel, channel receive will block
forever as no events are delivered.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>