fix: leave only a single IPv4/IPv6 address as kubelet's node IP
In Talos 0.13 we introduced a way to pick kubelet's node IP via machine
configuration, and later enforced that Talos picks up node IPs
automatically on boot (which got backported to 0.13.x).

The problem is that if the node has multiple IPv4/IPv6 addresses, Talos
builds broken configuration by default (with multiple `--node-ip` of the
same family) which won't let the kubelet start.

To fix the multiple addresses issue, one has to apply machine
configuration changes to pick just a single subnet.

To make matters worse, Talos 0.12.x won't accept any `nodeIP` machine
config setting (as it's not supported yet in 0.12), so upgrading to 0.13
leaves the node in a non-functional state.

The fix is to pick just a single IP of each family on Talos side and
print a warning to make sure user knows that one IP was ignored.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>