fix: filter down nameservers for docker-based cluster create
Filter the default list down to IPv4/IPv6 depending on the network
configuration. The problem is that the full list contains IPv6 resolvers
and these won't work on IPv4-only network (which is default for
`talosctl cluster create`).

This is not a problem for QEMU, as QEMU filters down the list for
built-in DHCP server already splitting it for DHCP4/DHCP6.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>