feat: enable accept_ra when IPv6 forwarding
Enables `accept_ra = 2` when IPv6 forwarding is enabled.

When IPv6 forwarding is enabled, the default `accept_ra = 1` no longer
functions.
This is intentional by the kernel developers, because routers generally should not
accept router advertisements (they supply their own).
However, in the case of a machine running Kubernetes, while IP
forwarding is enabled, the machine is still treated more as an end node
than a router.
It is common for a Kubernetes node to be configured via SLAAC and
therefore to expect to receive router advertisements, while at the same
time, IP forwarding must be enabled to handle container communication.

Fixes #3841

Signed-off-by: Seán C McCord <ulexus@gmail.com>