fix: refresh etcd certs on startup/join
This is more of a bandaid, rather than a real fix. As this should be
bacported to `release-1.0`, I tried to avoid doing big changes.

The race condition: controller correctly watches network state and
issues etcd certs as needed, but the service `etcd` writes down PKI
files from the resource just once early on startup. In this case there's
a chance that wrong PKI gets written to disk leaving etcd with
incomplete certs.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>