fix: clean up custom udev rules if the config is cleared
This fixes a case when udev rules are first created in the machine
config and then removed from the config.

As the file is on the overlayfs, it persists over reboots, so we need to
write it every time we boot Talos.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>
(cherry picked from commit 47d0e629d48930f6cb02dff32469bcb34440c73c)