docs: tag latest docs for search
This PR adds to our templates to inject a "latest" version metadata if
the doc is indeed our latest version. This will be coupled with an algolia
facet filter to only return "latest" version results in a separate PR
after I kick a recrawl of the site once this is merged.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>