fix: give up virtual IPs before the kubelet workloads are shut down
This introduces a shutdown lock basically which makes sure that whatever
needs to be terminated before kubelet shutdown sequence is initiated has
a chance to do so.

With the VIP, as the kubelet graceful shutdown terminates the API server
static pod, the VIP might be still on the node while the API server is
down. With this fix, VIP gets moved away from the node before the
kubelet graceful shutdown sequence starts.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>