docs: update releases shortcode in upgrade guide
This PR moves to using % instead of < for the release shortcode. As far
as I understand it, doing so tells hugo it's not raw html and thus the
headers get rendered to the sidebar properly. We may want to switch the
other uses as well, but I'm not sure it would make any difference one
way or another.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>