feat: add support for equinix metal events
This PR gets us closer to being a supported option on Equinix Metal. We
now look for a new kernel arg `em.events_url` and will send events there
based on the state of our install. These events are then presented to
the user in the GUI or via calls to the EM API.

*** Note: this is currently a WIP as the EM folks aren't yet ready to
test this out ***

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>