fix: correctly handle stopping services with reverse dependencies
This bug showed up with extension services: say we have a service
`ext-foo` which depends on service `cri`.

Service `ext-foo` will be started correctly only once `cri` is up.

But we should also stop `ext-foo` before `cri` is stopped, as otherwise
the dependency chain is broken. This PR fixes exactly that: once `cri`
is stopped, anything which depends on it should be stopped. We should
stop as well anything which depends on `ext-foo` (transitive
dependency).

In practical terms we use dependency on `cri` in extension service to
correctly stop/start extension services with `/var` filesystem
mount/unmount.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>