fix: run the 'post' stage of the service always
For most of the Talos service `post` stage does nothing, so it was never
properly noticed. FOr extension service, pre/post stages perform
mounting and unmounting of the overlayfs, so if post stage doesn't run
(if the runner can't be created), next time service is started, it won't
start as the post stage never ran.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>