fix: include Go primitive types into unstructured deepcopy
This code was written from JSON point of view, but
when YAML is unmarshaled,  we get more primitive Go types
as values, so why not include all of them?

This was showing as an error when applying a machine config e.g. for
kubelet extraArgs like:

```
shutdownGracePeriod: 0
```

Changing this to string fixes the problem, but it's not the best UX.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>
(cherry picked from commit 03ef62ad8bf10eccacbd490ebbcaceb0794f5378)