fix: ignore failures to dial wireguard client
This is required when running Talos in a container when host OS kernel
doesn't have support for Wireguard.

The failure is deferred so that controller actually fails when it needs
the wireguard client, but if wireguard is not used, controller continues
running.

This should fix errors when running Talos on Docker/WSL:

```
[talos] 2022/05/11 08:52:28 controller failed {"component": "controller-runtime", "controller": "network.LinkStatusController", "error": "error creating wireguard client: setsockopt: protocol not available"}
[talos] 2022/05/11 08:50:48 controller failed {"component": "controller-runtime", "controller": "kubespan.ManagerController", "error": "error creating wireguard client: setsockopt: protocol not available"}
[talos] 2022/05/11 08:50:32 controller failed {"component": "controller-runtime", "controller": "network.LinkSpecController", "error": "error creating wireguard client: setsockopt: protocol not available"}
```

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>
(cherry picked from commit 7fd1c80c3ee2ea33fe1aeb2e6cb21a3d0bc98537)