fix: reset certificate SANs on update
This affects both API server and Talos API cert SANs.

Before the fix, SANs accumulated changes over time, so even if the
hostname changes, old hostname is still kept in SANs. Even though it
shouldn't be a problem in general, it is confusing as after reboot list
will be reset back to expected value.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>