fix: detect WSL for `talosctl cluster create` on Docker
Docker on WSL seems to have same issue as Docker/OS X and
Docker/Windows: container IPs are not routable from the host, so we need
to apply same "magic" by using exposed ports.

WSL `talosctl` binary is built for Linux, so we need to do additional
checks for the platform.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>