fix: filter static pods correctly and optimize fetching
When we query kubelet API to populate the StaticPodStatuses, instead of checking for ownerReferences to be empty, we check the annotation "kubernetes.io/config.source" value so we avoid including standalone pods (that are regular pods but not part of a replicaset).

We also optimize their fetching by avoiding to unmarshal the fields we do not need.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>