fix: correctly unwrap errors when streaming
When message is sent via the proxy, `metadata.error` carries only string
representation which can't be unmarshalled back into an `error` which we
can match against. A similar fix was already done for "unary" responses,
but we missed the streaming case.

This fixes a spurious failure in integration tests when calling
`talosctl pcap --duration 1s`.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>